Eliminate n+1 queries and cache index pages (#43)
Uses a page caching gem to not hit the DB on the gatherings and books index. Also removes n+1 queries by using `length` instead of `count`. We could keep `count` (move work to DB) and use counter caches, but I don't want to do that yet, as we would  need to regenerate db.

Refreshes cache on updates.